+++ R/trajClusters.R
@@ -382,7 +382,7 @@ summary.trajClusters <- function(object, top_p = 3, ...) {
       median.j <- median(cl.medians[, j])
       
       deltas[, j] <- round(cl.medians[, j] - median.j, 4)
-      abs.deltas <- round(abs(cl.medians[, j] - median.j), 4)
+      abs.deltas <- abs(deltas[, j])
       n.unique <- length(unique(abs.deltas))
       #we wish to give the rank of 1 to ALL the groups which have the most extreme value, and the rank of 2 to ALL the groups that have the second most extreme value, etc.
       
@@ -405,30 +405,31 @@ summary.trajClusters <- function(object, top_p = 3, ...) {
     
     measure.names <- colnames(ranks)
     for( m in seq_len(length(measure.names))){
-      if(colnames(ranks[m]) == "m1"){ measure.names[m]  <- paste(colnames(ranks[m])," : max", sep = "")}
-      if(colnames(ranks[m]) == "m2"){ measure.names[m]  <- paste(colnames(ranks[m])," : min", sep = "")}
-      if(colnames(ranks[m]) == "m3"){ measure.names[m]  <- paste(colnames(ranks[m])," : range", sep = "")}
-      if(colnames(ranks[m]) == "m4"){ measure.names[m]  <- paste(colnames(ranks[m])," : mean", sep = "")}
-      if(colnames(ranks[m]) == "m5"){ measure.names[m]  <- paste(colnames(ranks[m])," : SD", sep = "")}
-      if(colnames(ranks[m]) == "m6"){ measure.names[m]  <- paste(colnames(ranks[m])," : slope", sep = "")}
-      if(colnames(ranks[m]) == "m7"){ measure.names[m]  <- paste(colnames(ranks[m])," : intercept", sep = "")}
-      if(colnames(ranks[m]) == "m8"){ measure.names[m]  <- paste(colnames(ranks[m])," : R²", sep = "")}
-      if(colnames(ranks[m]) == "m9"){ measure.names[m]  <- paste(colnames(ranks[m])," : int. rate", sep = "")}
-      if(colnames(ranks[m]) == "m10"){ measure.names[m]  <- paste(colnames(ranks[m])," : var. rate", sep = "")}
-      if(colnames(ranks[m]) == "m11"){ measure.names[m]  <- paste(colnames(ranks[m])," : contrast", sep = "")}
-      if(colnames(ranks[m]) == "m12"){ measure.names[m]  <- paste(colnames(ranks[m])," : tot var", sep = "")}
-      if(colnames(ranks[m]) == "m13"){ measure.names[m]  <- paste(colnames(ranks[m])," : spikiness", sep = "")}
-      if(colnames(ranks[m]) == "m14"){ measure.names[m]  <- paste(colnames(ranks[m])," : max f'", sep = "")}
-      if(colnames(ranks[m]) == "m15"){ measure.names[m]  <- paste(colnames(ranks[m])," : min f'", sep = "")}
-      if(colnames(ranks[m]) == "m16"){ measure.names[m]  <- paste(colnames(ranks[m])," : SD f'", sep = "")}
-      if(colnames(ranks[m]) == "m17"){ measure.names[m]  <- paste(colnames(ranks[m])," : f' var. rate", sep = "")}
-      if(colnames(ranks[m]) == "m18"){ measure.names[m]  <- paste(colnames(ranks[m])," : max f''", sep = "")}
-      if(colnames(ranks[m]) == "m19"){ measure.names[m]  <- paste(colnames(ranks[m])," : min f''", sep = "")}
-      if(colnames(ranks[m]) == "m20"){ measure.names[m]  <- paste(colnames(ranks[m])," : SD f''", sep = "")}
+      if(colnames(ranks[m]) == "m1"){ measure.names[m]  <- paste(colnames(ranks[m])," (max)", sep = "")}
+      if(colnames(ranks[m]) == "m2"){ measure.names[m]  <- paste(colnames(ranks[m])," (min)", sep = "")}
+      if(colnames(ranks[m]) == "m3"){ measure.names[m]  <- paste(colnames(ranks[m])," (range)", sep = "")}
+      if(colnames(ranks[m]) == "m4"){ measure.names[m]  <- paste(colnames(ranks[m]),": mean)", sep = "")}
+      if(colnames(ranks[m]) == "m5"){ measure.names[m]  <- paste(colnames(ranks[m])," (SD)", sep = "")}
+      if(colnames(ranks[m]) == "m6"){ measure.names[m]  <- paste(colnames(ranks[m])," (slope)", sep = "")}
+      if(colnames(ranks[m]) == "m7"){ measure.names[m]  <- paste(colnames(ranks[m])," (intercept)", sep = "")}
+      if(colnames(ranks[m]) == "m8"){ measure.names[m]  <- paste(colnames(ranks[m])," (R²)", sep = "")}
+      if(colnames(ranks[m]) == "m9"){ measure.names[m]  <- paste(colnames(ranks[m])," (int. rate)", sep = "")}
+      if(colnames(ranks[m]) == "m10"){ measure.names[m]  <- paste(colnames(ranks[m])," (var. rate)", sep = "")}
+      if(colnames(ranks[m]) == "m11"){ measure.names[m]  <- paste(colnames(ranks[m])," (contrast)", sep = "")}
+      if(colnames(ranks[m]) == "m12"){ measure.names[m]  <- paste(colnames(ranks[m])," (tot var)", sep = "")}
+      if(colnames(ranks[m]) == "m13"){ measure.names[m]  <- paste(colnames(ranks[m])," (spikiness)", sep = "")}
+      if(colnames(ranks[m]) == "m14"){ measure.names[m]  <- paste(colnames(ranks[m])," (max f')", sep = "")}
+      if(colnames(ranks[m]) == "m15"){ measure.names[m]  <- paste(colnames(ranks[m])," (min f')", sep = "")}
+      if(colnames(ranks[m]) == "m16"){ measure.names[m]  <- paste(colnames(ranks[m])," (SD f')", sep = "")}
+      if(colnames(ranks[m]) == "m17"){ measure.names[m]  <- paste(colnames(ranks[m])," (f' var. rate)", sep = "")}
+      if(colnames(ranks[m]) == "m18"){ measure.names[m]  <- paste(colnames(ranks[m])," (max f'')", sep = "")}
+      if(colnames(ranks[m]) == "m19"){ measure.names[m]  <- paste(colnames(ranks[m])," (min f'')", sep = "")}
+      if(colnames(ranks[m]) == "m20"){ measure.names[m]  <- paste(colnames(ranks[m])," (SD f'')", sep = "")}
     }
     
     for(i in seq_len(object$nclusters)){
-      w <- order(unlist(ranks[i, ]))[1:top_p]
+      # order the measures by increasing value of rank and, among measures of a given rank, by decreasing absolute value of delta
+      w <- order(unlist(ranks[i, ]), -unlist(abs(deltas[i, ])))[1:top_p]
       
       analysis[(i-1)*top_p + c(1:top_p), 1] <- i
       analysis[(i-1)*top_p + c(1:top_p), 2] <- measure.names[w]